test(picker): migrate trigger
